Pacman widget updated for pacman 3.3.
Pacman 3.3 returns one package on a line, without any other
information. So now we count lines, while before the list of packages
was provided on one line along with the number of updates - so we
didn't have to count them. Old code was commented out, and is waiting
for removal.