CPU Information widget included.
Widget returns speed and cache information for all available
CPUs/cores. It stores cpu speed in mhz and ghz in a table as well as
cpu cache in kb and mb. Values are retrieved using the CPU ID, i.e. we
would retrieve speed in ghz for CPU or core 1 with this format string:
${1 ghz}.