README updated in preparation for release 1.0.12.
Vicious tags from 1.0.12 will not be compatibile with awesome versions
prior to 3.4, tag 1.0.11 was the last one. Vicious was ported to the
new timer signals infrastructure and there is no backward
compatibility with hooks. In 1.0.12 even those C widgets that are
deprecated in awesome 3.4 (to be removed in 3.5) will not be
supported. Use awful.widget.progressbar and awful.widget.graph.