CHANGES file switched to 'oneline' short format
Now that vicious git repo is served trough a public web interface
there is no need for a full changelog. I considered removing it
completely, but it will have to wait - before the web interface
tarballed tags were requested much more than the repo was
cloned. Maybe that will change now, in which case the file should be
removed and stop wasting space.