contrib: Imported the rss widget, by olcc
It is a bit different from the overall vicious design, for now it is
in the contrib branch. We just gained some exposure and I expect more
people to send me stuff like this. It is still unclear weather vicious
will depend on third party libs (in which case we take an XML
parser). If not, maybe we will have a very basic html/xml parser as a
helper, or maybe neither and this widget will be introduced to master
(with a few more improvements I hope). It is too early to say, we just
got some users.