helpers: import capitalize
This helper will capitalize the first letter of every word in a given
string. It'll be useful for some widget string which look out of place
otherwise, like "rain, snow" (<- where did this come from?). But it
can also be useful for people that like to use this format, camel case
or simillar.