contrib: imported netcfg widget from Radu
Widget type by Radu provides active netcfg profiles in a table. In
contrib until we see what can be done with this, who wants to use it
and if data should be provided in other formats. We can't be certain
of the number of active profiles so a format string like $1|$2|$3 will
vary between sessions.