init: share timers when possible
We need time to give this a proper test with various usage scenarios;
multiple screens (and thus widget object instances), widgets
suspending and resuming and so on. Most benefits should come from
running on battery power (and if not suspending all widgets but
Battery it self on battery power), with less wake-ups.

Signed-off-by: Adrian C. (anrxc) <anrxc@sysphere.org>