gmail: fix subject when gmail feed is in one line
For some time now, gmail feeds are in a single line. The goal of this patch is to handle
both cases (single and multiline). It will find the text between title tags after the
first entry tag.

Since the first feed title is skiped with this regex, title regexes aren't needed anymore.

Signed-off-by: Adrian C. (anrxc) <anrxc@sysphere.org>